refactor: presentaional components file name changed to *Ui.tsx
