+++ web/src/shared/components/PageError.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 const PageError = () => {
-  return <div>PageError</div>;
+  return <div>404 Not found</div>;
 };
 
 export default PageError;
